Clean-up code and store port
